Restructuring directory, adding further tests
