+++ .gitignore
@@ -1,6 +1,6 @@
 # Logs
 logs
-report
+reports
 *.log
 npm-debug.log*
 yarn-debug.log*
@@ -52,23 +52,8 @@ jspm_packages/
 # Output of 'npm pack'
 *.tgz
 
-# Yarn Integrity file
-.yarn-integrity
-
 # Build
 dist
 
-# yarn v2
-<!-- .yarn/cache
-.yarn/unplugged
-.yarn/build-state.yml
-.yarn/install-state.gz
-.pnp.* -->
-
-.pnp.*
-.yarn/*
-!.yarn/patches
-!.yarn/plugins
-!.yarn/releases
-!.yarn/sdks
-!.yarn/versions
+# Test artifacts
+test-artifacts
