+++ package.json
@@ -67,10 +67,10 @@
     "@auto-it/omit-commits": "^11.0.4",
     "@auto-it/omit-release-notes": "^11.0.4",
     "@auto-it/released": "^11.0.4",
-    "@types/node": "^20.11.16",
-    "@typescript-eslint/eslint-plugin": "^6.20.0",
-    "@typescript-eslint/parser": "^6.20.0",
-    "@vitest/coverage-istanbul": "^1.2.2",
+    "@types/node": "^20.11.19",
+    "@typescript-eslint/eslint-plugin": "^7.0.1",
+    "@typescript-eslint/parser": "^7.0.1",
+    "@vitest/coverage-istanbul": "^1.3.0",
     "auto": "^11.0.4",
     "eslint": "^8.56.0",
     "eslint-config-airbnb-base": "^15.0.0",
@@ -78,10 +78,10 @@
     "eslint-import-resolver-typescript": "^3.6.1",
     "eslint-plugin-import": "^2.29.1",
     "eslint-plugin-prettier": "^5.1.3",
-    "prettier": "^3.2.4",
+    "prettier": "^3.2.5",
     "tempy": "^3.1.0",
-    "tsup": "^8.0.1",
+    "tsup": "^8.0.2",
     "typescript": "^5.3.3",
-    "vitest": "^1.2.2"
+    "vitest": "^1.3.0"
   }
 }
