--- reports/json-report.json
@@ -1,1 +1,0 @@
-{"numTotalTestSuites":7,"numPassedTestSuites":7,"numFailedTestSuites":0,"numPendingTestSuites":0,"numTotalTests":27,"numPassedTests":27,"numFailedTests":0,"numPendingTests":0,"numTodoTests":0,"startTime":1707353148868,"success":true,"testResults":[{"assertionResults":[{"ancestorTitles":["","convertPlainText"],"fullName":" convertPlainText should standardize line breaks by removing multiples return chatacters and tabs","status":"passed","title":"should standardize line breaks by removing multiples return chatacters and tabs","duration":7,"failureMessages":[]},{"ancestorTitles":["","convertPlainText"],"fullName":" convertPlainText should support skipping line break standardization","status":"passed","title":"should support skipping line break standardization","duration":1,"failureMessages":[]},{"ancestorTitles":["","convertPlainText"],"fullName":" convertPlainText should remove `\t`","status":"passed","title":"should remove `\t`","duration":0,"failureMessages":[]},{"ancestorTitles":["","convertPlainText"],"fullName":" convertPlainText should escape special characters","status":"passed","title":"should escape special characters","duration":2,"failureMessages":[]},{"ancestorTitles":["","convertPlainText"],"fullName":" convertPlainText should not inlining math equations by default","status":"passed","title":"should not inlining math equations by default","duration":1,"failureMessages":[]},{"ancestorTitles":["","convertPlainText"],"fullName":" convertPlainText should support inlining math equations through the $ latex character","status":"passed","title":"should support inlining math equations through the $ latex character","duration":0,"failureMessages":[]},{"ancestorTitles":["","convertImage"],"fullName":" convertImage should generate a latex image template","status":"passed","title":"should generate a latex image template","duration":6,"failureMessages":[]},{"ancestorTitles":["","convertImage"],"fullName":" convertImage should create a images directory if one does not exist","status":"passed","title":"should create a images directory if one does not exist","duration":5,"failureMessages":[]},{"ancestorTitles":["","convertImage"],"fullName":" convertImage should create a download the image using fetch and stream it to a file","status":"passed","title":"should create a download the image using fetch and stream it to a file","duration":1,"failureMessages":[]},{"ancestorTitles":["","convertImage"],"fullName":" convertImage should support setting the parent directory for the images directory","status":"passed","title":"should support setting the parent directory for the images directory","duration":2,"failureMessages":[]},{"ancestorTitles":["","convertImage"],"fullName":" convertImage should support autogenerated image filenames when saving the file","status":"passed","title":"should support autogenerated image filenames when saving the file","duration":1,"failureMessages":[]},{"ancestorTitles":["","convertImage"],"fullName":" convertImage should use a jpg as the default file extension if it cant be infered from the link only when autogenerating file names","status":"passed","title":"should use a jpg as the default file extension if it cant be infered from the link only when autogenerating file names","duration":0,"failureMessages":[]},{"ancestorTitles":["","convertImage"],"fullName":" convertImage should use the image's filename by default when saving the file","status":"passed","title":"should use the image's filename by default when saving the file","duration":1,"failureMessages":[]},{"ancestorTitles":["","convertImage"],"fullName":" convertImage should handle errors that occur while downloading the image","status":"passed","title":"should handle errors that occur while downloading the image","duration":0,"failureMessages":[]},{"ancestorTitles":["","convertImage"],"fullName":" convertImage should houtput logs if debug flag is passed and an error occurs during the file download","status":"passed","title":"should houtput logs if debug flag is passed and an error occurs during the file download","duration":13,"failureMessages":[]}],"startTime":1707353150705,"endTime":1707353150747,"status":"passed","message":"","name":"/home/flynn/Projects/personal/html-to-latex/test/unit/convert.mts"},{"assertionResults":[{"ancestorTitles":["","latexSpecialCharacter"],"fullName":" latexSpecialCharacter should escape special character: \"\\\"","status":"passed","title":"should escape special character: \"\\\"","duration":6,"failureMessages":[]},{"ancestorTitles":["","latexSpecialCharacter"],"fullName":" latexSpecialCharacter should escape special character: \"{\"","status":"passed","title":"should escape special character: \"{\"","duration":1,"failureMessages":[]},{"ancestorTitles":["","latexSpecialCharacter"],"fullName":" latexSpecialCharacter should escape special character: \"}\"","status":"passed","title":"should escape special character: \"}\"","duration":1,"failureMessages":[]},{"ancestorTitles":["","latexSpecialCharacter"],"fullName":" latexSpecialCharacter should escape special character: \"%\"","status":"passed","title":"should escape special character: \"%\"","duration":0,"failureMessages":[]},{"ancestorTitles":["","latexSpecialCharacter"],"fullName":" latexSpecialCharacter should escape special character: \"$\"","status":"passed","title":"should escape special character: \"$\"","duration":0,"failureMessages":[]},{"ancestorTitles":["","latexSpecialCharacter"],"fullName":" latexSpecialCharacter should escape special character: \"&\"","status":"passed","title":"should escape special character: \"&\"","duration":1,"failureMessages":[]},{"ancestorTitles":["","latexSpecialCharacter"],"fullName":" latexSpecialCharacter should escape special character: \"#\"","status":"passed","title":"should escape special character: \"#\"","duration":0,"failureMessages":[]},{"ancestorTitles":["","latexSpecialCharacter"],"fullName":" latexSpecialCharacter should escape special character: \"^\"","status":"passed","title":"should escape special character: \"^\"","duration":1,"failureMessages":[]},{"ancestorTitles":["","latexSpecialCharacter"],"fullName":" latexSpecialCharacter should escape special character: \"_\"","status":"passed","title":"should escape special character: \"_\"","duration":1,"failureMessages":[]},{"ancestorTitles":["","latexSpecialCharacter"],"fullName":" latexSpecialCharacter should escape special character: \"~\"","status":"passed","title":"should escape special character: \"~\"","duration":0,"failureMessages":[]},{"ancestorTitles":["","latexSpecialCharacter"],"fullName":" latexSpecialCharacter should escape special character: \"|\"","status":"passed","title":"should escape special character: \"|\"","duration":1,"failureMessages":[]}],"startTime":1707353150336,"endTime":1707353150349,"status":"passed","message":"","name":"/home/flynn/Projects/personal/html-to-latex/test/unit/templates.mts"},{"assertionResults":[{"ancestorTitles":["","exportFile"],"fullName":" exportFile should export latex file","status":"passed","title":"should export latex file","duration":10,"failureMessages":[]}],"startTime":1707353150132,"endTime":1707353150142,"status":"passed","message":"","name":"/home/flynn/Projects/personal/html-to-latex/test/unit/utils.mts"}]}
