--- reports/junit-report.xml
@@ -1,78 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<testsuites name="vitest tests" tests="27" failures="0" errors="0" time="1.907">
-    <testsuite name="test/unit/convert.mts" timestamp="2024-02-08T00:45:50.785Z" hostname="encom" tests="15" failures="0" errors="0" skipped="0" time="0.05">
-        <testcase classname="test/unit/convert.mts" name="convertPlainText &gt; should standardize line breaks by removing multiples return chatacters and tabs" time="0.007">
-        </testcase>
-        <testcase classname="test/unit/convert.mts" name="convertPlainText &gt; should support skipping line break standardization" time="0.001">
-        </testcase>
-        <testcase classname="test/unit/convert.mts" name="convertPlainText &gt; should remove `	`" time="0">
-        </testcase>
-        <testcase classname="test/unit/convert.mts" name="convertPlainText &gt; should escape special characters" time="0.002">
-        </testcase>
-        <testcase classname="test/unit/convert.mts" name="convertPlainText &gt; should not inlining math equations by default" time="0.001">
-        </testcase>
-        <testcase classname="test/unit/convert.mts" name="convertPlainText &gt; should support inlining math equations through the $ latex character" time="0">
-        </testcase>
-        <testcase classname="test/unit/convert.mts" name="convertImage &gt; should generate a latex image template" time="0.006">
-        </testcase>
-        <testcase classname="test/unit/convert.mts" name="convertImage &gt; should create a images directory if one does not exist" time="0.005">
-        </testcase>
-        <testcase classname="test/unit/convert.mts" name="convertImage &gt; should create a download the image using fetch and stream it to a file" time="0.001">
-        </testcase>
-        <testcase classname="test/unit/convert.mts" name="convertImage &gt; should support setting the parent directory for the images directory" time="0.002">
-        </testcase>
-        <testcase classname="test/unit/convert.mts" name="convertImage &gt; should support autogenerated image filenames when saving the file" time="0.001">
-        </testcase>
-        <testcase classname="test/unit/convert.mts" name="convertImage &gt; should use a jpg as the default file extension if it cant be infered from the link only when autogenerating file names" time="0">
-        </testcase>
-        <testcase classname="test/unit/convert.mts" name="convertImage &gt; should use the image&apos;s filename by default when saving the file" time="0.001">
-        </testcase>
-        <testcase classname="test/unit/convert.mts" name="convertImage &gt; should handle errors that occur while downloading the image" time="0">
-        </testcase>
-        <testcase classname="test/unit/convert.mts" name="convertImage &gt; should houtput logs if debug flag is passed and an error occurs during the file download" time="0.013">
-            <system-out>
-URL: http://test.com/image.jpg
-Error: Fetch Error
-    at [90m/home/flynn/Projects/personal/html-to-latex/[39mtest/unit/convert.mts:213:24
-    at [90mfile:///home/flynn/Projects/personal/html-to-latex/[39mnode_modules/[4m.pnpm[24m/@vitest+runner@1.2.2/node_modules/[4m@vitest[24m/runner/dist/index.js:128:14
-    at [90mfile:///home/flynn/Projects/personal/html-to-latex/[39mnode_modules/[4m.pnpm[24m/@vitest+runner@1.2.2/node_modules/[4m@vitest[24m/runner/dist/index.js:59:26
-    at runTest [90m(file:///home/flynn/Projects/personal/html-to-latex/[39mnode_modules/[4m.pnpm[24m/@vitest+runner@1.2.2/node_modules/[4m@vitest[24m/runner/dist/index.js:675:17[90m)[39m
-    at runSuite [90m(file:///home/flynn/Projects/personal/html-to-latex/[39mnode_modules/[4m.pnpm[24m/@vitest+runner@1.2.2/node_modules/[4m@vitest[24m/runner/dist/index.js:793:15[90m)[39m
-    at runSuite [90m(file:///home/flynn/Projects/personal/html-to-latex/[39mnode_modules/[4m.pnpm[24m/@vitest+runner@1.2.2/node_modules/[4m@vitest[24m/runner/dist/index.js:793:15[90m)[39m
-    at runFiles [90m(file:///home/flynn/Projects/personal/html-to-latex/[39mnode_modules/[4m.pnpm[24m/@vitest+runner@1.2.2/node_modules/[4m@vitest[24m/runner/dist/index.js:842:5[90m)[39m
-    at startTests [90m(file:///home/flynn/Projects/personal/html-to-latex/[39mnode_modules/[4m.pnpm[24m/@vitest+runner@1.2.2/node_modules/[4m@vitest[24m/runner/dist/index.js:851:3[90m)[39m
-    at [90mfile:///home/flynn/Projects/personal/html-to-latex/[39mnode_modules/[4m.pnpm[24m/vitest@1.2.2_@types+node@20.11.16/node_modules/[4mvitest[24m/dist/chunks/runtime-runBaseTests.QReNMrJA.js:114:7
-    at withEnv [90m(file:///home/flynn/Projects/personal/html-to-latex/[39mnode_modules/[4m.pnpm[24m/vitest@1.2.2_@types+node@20.11.16/node_modules/[4mvitest[24m/dist/chunks/runtime-runBaseTests.QReNMrJA.js:82:5[90m)[39m
-
-            </system-out>
-        </testcase>
-    </testsuite>
-    <testsuite name="test/unit/templates.mts" timestamp="2024-02-08T00:45:50.797Z" hostname="encom" tests="11" failures="0" errors="0" skipped="0" time="0.016">
-        <testcase classname="test/unit/templates.mts" name="latexSpecialCharacter &gt; should escape special character: &quot;\&quot;" time="0.006">
-        </testcase>
-        <testcase classname="test/unit/templates.mts" name="latexSpecialCharacter &gt; should escape special character: &quot;{&quot;" time="0.001">
-        </testcase>
-        <testcase classname="test/unit/templates.mts" name="latexSpecialCharacter &gt; should escape special character: &quot;}&quot;" time="0.001">
-        </testcase>
-        <testcase classname="test/unit/templates.mts" name="latexSpecialCharacter &gt; should escape special character: &quot;%&quot;" time="0">
-        </testcase>
-        <testcase classname="test/unit/templates.mts" name="latexSpecialCharacter &gt; should escape special character: &quot;$&quot;" time="0">
-        </testcase>
-        <testcase classname="test/unit/templates.mts" name="latexSpecialCharacter &gt; should escape special character: &quot;&amp;&quot;" time="0.001">
-        </testcase>
-        <testcase classname="test/unit/templates.mts" name="latexSpecialCharacter &gt; should escape special character: &quot;#&quot;" time="0">
-        </testcase>
-        <testcase classname="test/unit/templates.mts" name="latexSpecialCharacter &gt; should escape special character: &quot;^&quot;" time="0.001">
-        </testcase>
-        <testcase classname="test/unit/templates.mts" name="latexSpecialCharacter &gt; should escape special character: &quot;_&quot;" time="0.001">
-        </testcase>
-        <testcase classname="test/unit/templates.mts" name="latexSpecialCharacter &gt; should escape special character: &quot;~&quot;" time="0">
-        </testcase>
-        <testcase classname="test/unit/templates.mts" name="latexSpecialCharacter &gt; should escape special character: &quot;|&quot;" time="0.001">
-        </testcase>
-    </testsuite>
-    <testsuite name="test/unit/utils.mts" timestamp="2024-02-08T00:45:50.805Z" hostname="encom" tests="1" failures="0" errors="0" skipped="0" time="0.012">
-        <testcase classname="test/unit/utils.mts" name="exportFile &gt; should export latex file" time="0.01">
-        </testcase>
-    </testsuite>
-</testsuites>
