+++ src/index.mts
@@ -1,2 +1,2 @@
-export { convertText, convertFile } from './convert.mts';
-export type { ConvertOptions, ConvertFileOptions } from './convert.mts';
+export { convertText, convertFile } from './helpers/index.mts';
+export type { ConvertOptions, ConvertFileOptions } from './types.mts';
