--- test/assets/3/index.html
@@ -1,3 +1,0 @@
-Three concentric metal shells \(A,{\rm{ }}B\) and \(C\) of respective radii \(a,{\rm{ }}b\) and \(c\) \((a < b < c)\) have surface charge densities \( + \sigma , - \sigma \) and \( + \sigma \) respectively. The potential of shell \({\rm{B}}\) is: <br> <img src="images/image1.png">
-
-Net potential at \({\rm{B}}\) is due to superposition of plotential due to all shells.<br>\({V_B} = {V_{B(I)}} + {V_{B(II)}} + {V_{B(III)}}\)<br>\( = \cfrac{{K4\pi {a^2}\sigma }}{b} + \cfrac{{K4\pi {b^2}( - \sigma )}}{b} + \cfrac{{K4\pi {c^2}(\sigma )}}{c}\)<br>Required potential of b<br>\( = \cfrac{{\sigma {a^2}}}{{{\varepsilon _0}b}} - \cfrac{{\sigma {b^2}}}{{{\varepsilon _0}b}} + \cfrac{{\sigma {c^2}}}{{{\varepsilon _0}c}}\)<br>\( = \cfrac{\sigma }{{{\varepsilon _0}}}\left( {\cfrac{{{a^2} - {b^2}}}{b} + c} \right)\)
