--- test/assets/3/output.tex
@@ -1,28 +1,0 @@
-\documentclass{article}
-
-\usepackage{amsmath}
-\usepackage{graphicx}
-
-
-\begin{document}
-
-Three concentric metal shells \(A,{\rm{ }}B\) and \(C\) of respective radii \(a,{\rm{ }}b\) and \(c\) \((a < b < c)\) have surface charge densities \( + \sigma , - \sigma \) and \( + \sigma \) respectively. The potential of shell \({\rm{B}}\) is:
-
-\begin{center}
-	\includegraphics{images/image2.png}
-\end{center}
-
-Net potential at \({\rm{B}}\) is due to superposition of plotential due to all shells.
-
-\({V_B} = {V_{B(I)}} + {V_{B(II)}} + {V_{B(III)}}\)
-
-\( = \cfrac{{K4\pi {a^2}\sigma }}{b} + \cfrac{{K4\pi {b^2}( - \sigma )}}{b} + \cfrac{{K4\pi {c^2}(\sigma )}}{c}\)
-
-Required potential of b
-
-\( = \cfrac{{\sigma {a^2}}}{{{\varepsilon _0}b}} - \cfrac{{\sigma {b^2}}}{{{\varepsilon _0}b}} + \cfrac{{\sigma {c^2}}}{{{\varepsilon _0}c}}\)
-
-\( = \cfrac{\sigma }{{{\varepsilon _0}}}\left( {\cfrac{{{a^2} - {b^2}}}{b} + c} \right)\)
-
-
-\end{document}
