+++ test/unit/helpers/convert-ordered-list.mts
@@ -1,0 +1,36 @@
+import { parseFragment } from 'parse5';
+import { expect, describe, test as it, vi } from 'vitest';
+import * as Template from '../../../src/templates.mts';
+import { convertOrderedLists } from '../../../src/helpers/convert-ordered-list.mts';
+import * as convertListUtils from '../../../src/helpers/convert-list.mts';
+import { ElementNode } from '../../../src/types.mts';
+
+describe('convertOrderedLists', () => {
+  it('should parse out the li element children from the node using the convertList helper', async () => {
+    const convertListSpy = vi.spyOn(convertListUtils, 'convertList');
+
+    const node = parseFragment('<ol><li>one</li><li>two</li><span>three</span></ol>');
+
+    await convertOrderedLists(node.childNodes[0] as ElementNode);
+
+    expect(convertListSpy).toHaveBeenCalledWith(node.childNodes[0], {});
+  });
+
+  it('should use the enumerate template for each of the converted li blocks', async () => {
+    const templateSpy = vi.spyOn(Template, 'enumerate');
+
+    const node = parseFragment('<ol><li>one</li><li>two</li><span>three</span></ol>');
+
+    await convertOrderedLists(node.childNodes[0] as ElementNode);
+
+    expect(templateSpy).toHaveBeenCalledWith;
+  });
+
+  it('should convert the list elements using enumerate environment', async () => {
+    const node = parseFragment('<ol><li>one</li><li>two</li><span>three</span></ol>');
+
+    const output = await convertOrderedLists(node.childNodes[0] as ElementNode);
+
+    expect(output).toMatch('\\begin{enumerate}\n\t\\item one\n\t\\item two\n\\end{enumerate}');
+  });
+});
