+++ tsconfig.json
@@ -22,6 +22,6 @@
     "noUnusedParameters": true,
     "noImplicitReturns": true,
 
-    "types": ["jest", "node"],
-  },
+    "types": ["jest", "node"]
+  }
 }
